psoib: Hook into __morecore to retrieve "safe" heap regions.
psoib_mregion_cache() can cache only regions which can not be
free()'d by the user without the caches knowledge. We have to either,
reliable detect a free (and invalidate the cache) or intercept the
free (not giving the mem back to the kernel). If we can't, we should
not cache. For now, we assume only the region between NULL and sbrk()
to be "safe" and do not cache anonymous mmap'ed mem. At the same time,
we set mallopt(M_MMAP_MAX, 0) with the hope, mmap is not or rarely
used.