debug: Enable custom string formats for debug writes.
Instead of using a fixed format for every debug write, `pscom_dwrite()`
now expects the format to be specified per call.

Additionally, instead of writing everything to the global debug file
(that is commonly used for debugging purpose as the name suggests),
pscom_dwrite() now expects the explicit definition of the output stream
to be used.