shutdown: Handle generated net requests
This commit addresses the (abnormal) case in the connection shutdown
protocol where unexpected messages still arrive after the closing of
a connection has been actively triggered. Such cases may occur after
an MPI_Abort() call and/or an internal MPI error.

The simple solution of this commit is to repeatedly check and empty
the respective request queue before the final teardown.