shutdown: Add timeout for pscom_sock_close()
It turned out that the while(1) loop, which checks for all connections
being closed for a synchronized shutdown in the at-exit handler, can
actually run into a deadlock if other processes do no further progress
on this. (In #11, we have a reproducer for this issue, where MPI_ABORT
is called by a single process and not collectively.)

This commit adds a simple timeout mechanism for this that terminates
the loop after PSP_SHUTDOWN_TIMEOUT seconds with no progress.
This is implemented by counting the number of the still non-closed
connections and then by checking  whether this number is further
decreasing. If this is the case, a new start time for the timeout is
taken (timer restart).

If the PSP_SHUTDOWN_TIMEOUT environment variable is not set or set
to 0 (default), the timeout is infinite.