shutdown: Add PSP_DEBUG_BYE_MSG for debugging
As the notorious "Byee." message is the very last debug message that
pscom emits, it might be useful (especially for large system runs)
to have only this message be printed at the end to check whether all
processes have successfully passed the pscom's exit handler.

The output of only this one message can be achieved by setting the
new PSP_DEBUG_BYE_MSG environment variable to 1.