pscom: Fix warnings about unknown attribute `__nonstring__`
When building (e.g., via pscom "allin" for psmpi) with the icc or nvc
compiler, a lot of warnings about unknown an attribute `__nonstring__`
flood the build log. This commit fixes this issue by checking with the
`__has_attribute` macro whether this attribute is actually supported.

This reverts 20726a893 since the check also covers older gcc versions
that did not suppport `__nonstring__`.