cuda: Only copy back from stage memory if request was successful
If a receive request fails (e.g., due to a connection shutdown before
the request can be satisfied) or is canceled explicitly, the copy-back
operation for unstaging the receive buffer is unnecessary -- and can
probably also be problematic in certain cases.

This commit therefore adds an according check against the error and the
cancel state before copying back data from device to host memory.