rma: Add memory region and remote key data structures.
This patch introduces the concept of memory regions and remote keys by
adding the respective data structures to the public and private
interfaces.

Additionally, function pointers for RMA-related memory registration and
remote key handling are added to the `pscom_arch_sock` and `pscom_con_t`
structures. In doing so, a clearer separation between rendezvous-related
and RMA-related function is achieved.

Finally, the plugin and unit tests are adapted to the changes described
above.