cuda: Avoid staging of CUDA buffers for on-demand connections.
For on-demand connections `is_gpu_aware` is set to 0 in
`pscom_con_create()`. When the first send/recv occurs on such a
connection, the buffer is staged because `_pscom_buffer_needs_staging()`
assumes the connection is not GPU-aware (i.e., `is_gpu_aware` is set to
zero). When connection is finally set up for each plugin, `is_gpu_aware`
is updated and subsequent sends/recvs do not use staging (in case of
GPU-aware connections).

To avoid the usage of the staging buffer upon the first recv/send,
this patch moves buffer staging just before IO has been started.