listener: Add new API to suspend/resume listening without closing the fd.
This commit adds two new functions to the pscom API:

pscom_suspend_listen: Stop listening for new connections on the socket's
port by deactivating the ufd entry in the progress engine. This function
does *not* close the fd or return the port to the OS.

pscom_resume_listen: Start listening on the socket's port by adding the
udf entry back to the progress engine. This function must be used in pair
with pscom_suspend_listen.

The define switch PSCOM_HAS_LISTEN_SUSPEND_RESUME indicates if the new API
functions are available for pscom users.