utest/io: Fix jumps depending on uninitialized data.
When using valgrind, some of the tests in `test_io.c` reported
warnings about conditional jumps depending on uninitialized data.
The reason was the use of requests with header data that had not
been initialized (as it is not actually being used as part of
the test).

This commit fixes this issue by initializing explicitly the related
requests (e.g., by calling `pscom_req_prepare_send()`) and with this
also their header-related fields before using them.