ufd: Tag ufd elements for their thread-safe handling.
In the threaded case, each thread makes a local copy of the global ufd
object and its arrays, and then calls the `poll()` function with this
copy. The (local) results, reported as ufd events, are then processed
concurrently in a lock-protected loop in which the lock can, however,
be temporarily released again. To provide thread safety during this
concurrent processing, changes to the global ufd object were previously
detected by the threads by pointer comparisons between global and local
array entries of the ufd objects.

However, as pointer addresses can be released and then re-used, their
values are practically not well suited for indicating updates here.

This commit therefore introduces the notion of "unique tags" instead,
implemented as sequence numbers that are assigned and updated for every
write access to a global ufd array entry. In doing so, each respective
thread is only allowed to work on an item if the tag of the local copy
is still the same as the tag of the global counterpart.

Since this mechanism is only needed in the threaded case, the tags are
only assigned and the corresponding array to store them is only created
in that case. Therefore, the non-threaded case should not be affected
by the changes in this commit.

However, the changes in this commit also required an adjustment to the
ufd related unit tests regarding the initialization and cleanup of the
global ufd object.