utest: Reset threaded mode in teardown_dummy_con.
If a unit test fails in threaded mode, the test suite can run into
a deadlock when the dummy connections are closed and the threaded
mode with the related locks are still enabled.

This commit is a simple fix (at least for the time being) that avoids
such deadlocks by resetting `pscom.threaded` explicitly to zero in
`teardown_dummy_con()`.

However, perhaps a more sophisticated infrastructure for tests in
threaded mode should be considered and implemented for the future.