listener: Delete ufd_info from list only when active cnt drops to 0.
If the active counter of the listener drops to 0, delete the
ufd_info only once from the list (at the time when the counter
drops to 0). If pscom_listener_active_dec is called in suspend
mode later on, e.g., during cleanup, the ufd_info is not deleted
from the list again.